requirements.txt
pandas
boto3
requests
python-dotenv
pytest